% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{tidypairs}
\alias{tidypairs}
\title{Generate data for pairwise scatterplot matrix.}
\usage{
tidypairs(data, xy_vars, z_vars = NULL, mirror = F, diagonal = F)
}
\arguments{
\item{data}{Data frame (in "wide" format, i.e. must contain a column for each
xy_vars and z_vars).}

\item{xy_vars}{Character vector indicating variables to be paired (these
would be the x and y dimensions if the result is used for a scatterplot
matrix).}

\item{z_vars}{Optional character vector indicating additional variables to
retain for each record.}

\item{mirror}{Logical indicating whether to include data for both upper and
lower triangles of the matrix (default is FALSE).}

\item{diagonal}{Logical indicating whether to include data for the diagonal
of the matrix, in which the x and y variables match (default is FALSE).}
}
\value{
A restructured data frame, with columns `x_var`, `y_var`, `x_value`,
  and `y_value`, as well as a column for each of `z_vars`.
}
\description{
Convert a wide dataset into a format that can be fed to ggplot to create a
scatterplot matrix. The dataset is restructured into a format with columns
for the x and y values, and columns for the x and y facets.
}
\examples{
d <- tidypairs(iris, xy_vars = names(iris)[1:4], z_vars = names(iris)[5])
library(ggplot2)
ggplot(d, aes(x_value, y_value, color = Species)) +
  facet_grid(y_var ~ x_var, scales = "free") +
  geom_point()
}
